+++ src/OpenTelemetry.Instrumentation.AWSLambda/AWSLambdaWrapper.cs
@@ -17,7 +17,7 @@ public static class AWSLambdaWrapper
 {
     internal const string ActivitySourceName = "OpenTelemetry.Instrumentation.AWSLambda";
 
-    private static readonly ActivitySource AWSLambdaActivitySource = new(ActivitySourceName, typeof(AWSLambdaWrapper).Assembly.GetPackageVersion());
+    private static readonly Lazy<ActivitySource> AWSLambdaActivitySource = new(CreateActivitySource);
 
     private static bool isColdStart = true;
 
@@ -172,7 +172,7 @@ public static class AWSLambdaWrapper
 
         // We assume that functionTags and httpTags have no intersection.
         var activityName = AWSLambdaUtils.GetFunctionName(context) ?? "AWS Lambda Invoke";
-        var activity = AWSLambdaActivitySource.StartActivity(activityName, ActivityKind.Server, parentContext, functionTags.Concat(httpTags)!, links);
+        var activity = AWSLambdaActivitySource.Value.StartActivity(activityName, ActivityKind.Server, parentContext, functionTags.Concat(httpTags)!, links);
 
         return activity;
     }
@@ -254,5 +254,19 @@ public static class AWSLambdaWrapper
         {
             OnFunctionStop(activity, tracerProvider);
         }
+    }
+
+    private static ActivitySource CreateActivitySource()
+    {
+        var assembly = typeof(AWSLambdaWrapper).Assembly;
+        var version = assembly.GetPackageVersion();
+
+        var activitySourceOptions = new ActivitySourceOptions(ActivitySourceName)
+        {
+            TelemetrySchemaUrl = AWSSemanticConventions.SchemaUrl,
+            Version = version,
+        };
+
+        return new(activitySourceOptions);
     }
 }