+++ src/OpenTelemetry.Instrumentation.AWSLambda/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Fix `faas.max_memory` overflow for Lambda memory settings at or above 2048 MB.
   ([#4125](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4125))
 
+* Add instrumentation scope version and schema URL to traces.
+  ([#4063](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4063))
+
 ## 1.15.0
 
 Released 2026-Jan-21